RAG_QAbot/requirements.txt
ipykernel
langchain_community
langchain
streamlit
langchain-chroma
langchain-core
python-dotenv
langchain_ollama
pypdf